Rename com.hpe.sbt to com.github.sbt.pullrequestvalidator
