+++ src/main/scala/com/hpe/sbt/package.scala
@@ -1,0 +1,6 @@
+package com.hpe
+
+package object sbt {
+  @deprecated("com.github.sbt.pullrequestvalidator.ValidatePullRequest", "2.0.0")
+  val ValidatePullRequest = com.github.sbt.pullrequestvalidator.ValidatePullRequest
+}
